Moving data_store_conduit.
