+++ include/lbann/data_ingestion/readers/data_reader_image.hpp
@@ -29,9 +29,9 @@
 #ifndef IMAGE_DATA_READER_HPP
 #define IMAGE_DATA_READER_HPP
 
+#include "lbann/data_ingestion/data_store_conduit.hpp"
 #include "lbann/data_ingestion/readers/data_reader.hpp"
 #include "lbann/data_ingestion/readers/sample_list.hpp"
-#include "lbann/data_store/data_store_conduit.hpp"
 
 namespace lbann {
 class image_data_reader : public generic_data_reader
