+++ src/data_ingestion/CMakeLists.txt
@@ -23,6 +23,10 @@
 ## implied. See the License for the specific language governing
 ## permissions and limitations under the license.
 ################################################################################
+# Add the source files for this directory
+set_full_path(THIS_DIR_SOURCES
+  data_store_conduit.cpp
+)
 
 # Add the subdirectories
 add_subdirectory(infrastructure)
