+++ src/data_ingestion/readers/data_reader_smiles.cpp
@@ -27,10 +27,10 @@
 
 #include "lbann/data_ingestion/readers/data_reader_smiles.hpp"
 #include "lbann/comm_impl.hpp"
+#include "lbann/data_ingestion/data_store_conduit.hpp"
 #include "lbann/data_ingestion/readers/data_reader_sample_list_impl.hpp"
 #include "lbann/data_ingestion/readers/sample_list_impl.hpp"
 #include "lbann/data_ingestion/readers/sample_list_open_files_impl.hpp"
-#include "lbann/data_store/data_store_conduit.hpp"
 #include "lbann/utils/argument_parser.hpp"
 #include "lbann/utils/commify.hpp"
 #include "lbann/utils/file_utils.hpp"
